+++ docs/source/TimerMechanism.rst
@@ -1,0 +1,7 @@
+TimerMechanism
+==============
+
+.. automodule:: psyneulink.library.components.mechanisms.processing.integrator.timermechanism
+   :members:
+   :private-members:
+   :exclude-members: random, Parameters
