+++ docs/source/TransferMechanisms.rst
@@ -17,4 +17,3 @@ Transfer Mechanisms
    KWTAMechanism
    KohonenMechanism
    ContrastiveHebbianMechanism
-   TimerMechanism
