+++ psyneulink/library/components/mechanisms/processing/integrator/__init__.py
@@ -1,8 +1,11 @@
 from . import ddm
 from . import episodicmemorymechanism
+from . import timermechanism
 
 from .ddm import *
 from .episodicmemorymechanism import *
+from .timermechanism import *
 
 __all__ = list(ddm.__all__)
 __all__.extend(episodicmemorymechanism.__all__)
+__all__.extend(timermechanism.__all__)
