+++ psyneulink/library/components/mechanisms/processing/transfer/recurrenttransfermechanism.py
@@ -6,7 +6,7 @@
 # See the License for the specific language governing permissions and limitations under the License.
 
 # NOTES:
-#  * NOW THAT NOISE AND INTEGRATION_RATE ARE PROPRETIES THAT DIRECTLY REFERERENCE integrator_function,
+#  * NOW THAT NOISE AND INTEGRATION_RATE ARE PROPERTIES THAT DIRECTLY REFERERENCE integrator_function,
 #      SHOULD THEY NOW BE VALIDATED ONLY THERE (AND NOT IN TransferMechanism)??
 #  * ARE THOSE THE ONLY TWO integrator PARAMS THAT SHOULD BE PROPERTIES??
